LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/after-scrollable-range-shrinkage-001-expected.txt


FAIL Scroll anchoring properly works after scrollable range shrinkage assert_equals: expected 608 but got 626
